CLI: log to stderr instead of stdout
Since the `ls` command produces machine-readable output, we need to ensure
that any log messages produced do not corrupt that output.

Logging to stderr is also conventional; Python's `StreamHandler` even
defaults to it.

This breaks the `import` command tests, because previously they were looking
at the last log message. I don't think we should be testing log messages, so
add a proper `print` to this command. This also makes it consistent with the
`create` command.