Fix an issue of releasing lock for rq export job when the worker subprocess is killed (#8721)
The main problem fixed by this PR is as follows:
In the previous implementation, "long" locks were used when exporting a
resource or deleting an export cache.
If the export process was killed (e.g., by the OOM killer with 9
signal), the acquired lock was not released and remained active until
the auto-release timeout expired (e.g., 4 hours). A subsequent user
request to export a dataset could not acquire the lock, causing the job
to be scheduled for execution after 60 seconds (default value). When the
scheduled job ran again, it still could not acquire the lock, and the
entire process was repeated. Additionally, if a user initiated the
export process after the job was marked as scheduled, they were unable
to re-initiate the process and received an error because the RQ job
status was not set and handled correctly (it was remaining `STARTED`).

One more found and fixed problem is that 2 users that have rights to
export a resource could not make export in parallel (with the same
options like format, save_images) and one of them received a
`LockNotAvailableError` error.

---------

Co-authored-by: Maxim Zhiltsov <zhiltsov.max35@gmail.com>