Simplify the code formatting script
After #8611 and #8866, it's no longer necessary to apply the formatters to
each module separately. So don't!

In addition, update the black exclusion list in `pyproject.toml` to only contain
files that were _not_ listed in `format_python_code.sh`. This way, formatting
will automatically be enforced for all new files.

For a few of the files reformatting only creates a small diff, so don't even
bother excluding them.