Change match_empty_frames to empty_is_annotated (#8888)
Improves the implementation of the `match_empty_frames` quality setting
to work in cases when only GT or DS annotations are empty. This allows
to use precision as the primary metric in cases when all frames are
required to have at least 1 annotation.

The previous implementation only affected matching empty annotations.
The updated variant also counts any empty frames in `ds_count` and
`gt_count` so that mismatched empty annotations also included in
denominators of metrics.

Example: 

| annotations per frame | old accuracy | new accuracy | old precision |
new precision |
| - | - | - | - | - |
| ds: [empty, 1 valid]; gt: [empty, 1 valid] | 2/2 | 2/2 | 2/2 | 2/2 |
| ds: [1 extra, 1 valid]; gt: [empty, 1 valid] | 1/2 (only matches) |
1/3 (empty != extra) | 1/2 | 1/2 |
| ds: [empty, 1 valid]; gt: [1 miss, 1 valid] | 1/2 (only matches) | 1/3
(empty != miss) | 1/1 (only matches) | 1/2 |

So, it allowed undesirable situations in which 1 and the only correct
annotation in a job could be counted as 100% of precision. The updated
option prevents this.