Merge commit from fork
Tracker functions have to preserve the tracking state by returning it to the
client after a call and then accepting it back on the next call. To do this,
a tracker function must be able to serialize its internal state into a JSON
value and deserialize it back. It can be tempting to use a generic
serialization library to do this. In fact, both of the functions in this
repo do just that (specifically, they use jsonpickle).

The problem with such libraries (and jsonpickle in particular) is that they
are not safe to use on untrusted data, as it may contain instructions to run
arbitrary code. And since the state data supplied by the client is passed
directly to the function, the client can execute arbitrary code in the
function container.

This could be fixed on the individual function level, by using safer
serialization methods, but that would mean a duplication of effort, as each
function would have to be secured individually. It would also do nothing
to protect customers with their own custom tracker functions, which might
suffer from the same problem.

Instead, we can fix this on the CVAT server level, once and for all. We can
append a MAC to each state before passing it from the function to the
client, and verify the MAC when going in the opposite direction. This
ensures that the tracker only deserializes states that it itself has
serialized. This works transparently as far as the UI is concerned, since
it does not examine or modify the state it receives from the function.