Add skeleton support for native functions
All of the support code in the auto-annotation SDK layer is already there,
we just need to send sublabel information when creating a native function,
and validate it when running an agent.

In addition, force keypoint sublabel types to be "points", since this is how
the UI creates them for tasks/projects.