Add attribute support for auto-annotation functions (#9090)
Remove one of the long-standing limitations on auto-annotation functions
by adding the necessary validation and remapping logic to support
attribute specifications and values. Add a utility module for attributes
with functionality I needed, but felt didn't belong in the
auto-annotation layer.

Adds the necessary code to support using functions with attributes via
agents, as well. I will submit the necesssary server-side code will be
submitted to the private repository later; until that is merged,
attempts to create native functions with attributes will be rejected.