Merge the `_xxx_request` helper methods in unit tests
There are two different versions of these methods, in `ApiTestBase` and in
`_LambdaTestCaseBase`. Since the latter inherits from the former, Pylint
warns that the overridden methods are incompatible (and they are!). To fix
this, merge these two sets of methods into one, with all needed
capabilities.

Some notes on the decisions I made WRT the merged methods:

1. In `_post_request` I emulated the `query_params` parameter from Django
   5.1.

2. In `_get_request` I renamed `data` to `query_params` for consistency.

3. In `_patch_request` I decided not to add a `query_params` parameter for
   now, since it wasn't needed. While some lambda tests called
   `_patch_request` with an org ID, this parameter was unnecessary, so I
   just removed it.

4. I set the default format to "json" everywhere, since CVAT's API is mostly
   JSON-based.