lambda_manager: output detection results in the LabeledData format (#9285)
With Nuclio detector functions, we have code on the server that converts
the output of the function into CVAT API format annotations that is used
when the user runs a function on the whole task. But we also have
roughly the same code on the client that is used when the user runs a
function on a frame. This is unnecessary code duplication, and it means
that there could be subtle differences with how the function output is
interpreted in batch vs interactive requests.

In addition, this creates an annoying situation with agents, which
already create annotations in the CVAT API format. To make agents work
with per-frame annotation (which I would like to do), we'd have to
convert those annotations to the Nuclio format, send them to the UI, and
have the UI convert them again, which is just silly.

Therefore, in this patch I make the server do the conversion for both
batch and interactive requests. Note that the UI doesn't _quite_ expect
the results in the API format (it wants `ObjectState`s), so a small
amount of conversion logic in the UI remains; but it's much simpler than
before, since the server does the bulk of the work. In the future it
would be nice to refactor this to reuse the logic from
`Collection.import`, but I don't feel confident enough in the frontend
codebase to do it at the moment.

This patch also moves the responsibility of handling the "Convert masks
to polygons" parameter for interactive requests from the client to the
server (for batch requests, the server was already handling it).