Switch to deploying GitHub Pages via an action
Currently, we're using the original deployment method of pushing to the
`gh-pages` branch. Since the actual content is generated via a workflow,
we can skip the pushing, and just publish directly to Pages from the workflow.

This has the following benefits:

* It should be a bit faster, since we're skipping a step.
* It makes the repository lighter (provided we delete the `gh-pages` branch).
* It lets you preview what pages will be deployed after a PR,
  since they are published as a workflow artifact.
* It lets us avoid using a 3rd-party action.

Note that `actions/deploy-pages` has no equivalent of the `cname` parameter.
Instead, the domain name is set in GitHub's repository settings.