SDK: Deprecate server URL scheme autodetection (#9384)
This feature is fundamentally insecure. If a user makes use of this,
then a MitM could trivially block the HTTPS connection attempt, causing
the SDK to downgrade to HTTP. Even in the absence of an active attacker,
a transient error could cause a downgrade, leading to request details
and credentials being exposed in cleartext.

In addition, this feature creates a usability problem, because if the
autodetection fails, the user is given no information about the cause.
This could be fixed, but given the main issue, I don't think it's
worthwhile.

I don't think we can just remove this feature immediately, given that
the SDK has worked this way for a long time, and there are certainly
people relying on it. Therefore, in this patch I just add a big warning,
and recommend specifying the scheme explicitly. In a future release, we
can drop the autodetection.