Upgrade remark to the latest version (#9394)
The new version seems to be better at parsing Markdown and emits more
useful warnings (which I fixed). The new warnings are mostly of the
following categories:

1. Inconsistent table formatting.
2. Inconsistent list numbering.
3. Overlong lines.
4. Missing quote markers.

WRT tables, I converted a few of them to the compact style, because I
can't be bothered to align rows that are >100 characters long, and I
don't think anyone else should have to.