Documentation website: Change links for logo and About in the header (#9437)
Our marketing team asked to change the **About** link in the header. It
should lead to the main website.
By the same logic, I changed the logo link that previously led to the
**About** page as well. Also, I hid the same link in the footer.

I deleted the HTML file for the About page. The 404 page had a link to it too,
so I changed this HTML as well.