Add alt texts for images (#9499)
Google says [1] alt text is the most important metadata for images.
We had 300 images without it in the documentation, and this PR adds
all of the missing alt texts. Hopefully, it helps our documentation website
in search engine rankings.

I have also fixed a couple of mistypes and double empty lines in this
PR.

[1] https://developers.google.com/search/docs/appearance/google-images#descriptive-alt-text%20descriptive-titles-captions-filenames