SDK: Refactor to better support other function types (#9530)
Currently, we have `DetectionFunction` hardcoded in several places where
conceptually any type of function should be allowed. Introduce a generic
`AutoAnnotationFunction` protocol and use it where appropriate.

In addition, add a specific error message from when a CLI loads a
function that has no `spec` attribute (suggesting it's not a function at
all).