Upgrade pylint to 3.x (#9498)
Additionally, change the workflow to scan the entire codebase. A change
in one file can cause a pylint warning in another, so this is the only
way to avoid committing code with warnings.

I disabled several new warnings, since there were too many instances of
those to fix in one sittings. They can be fixed separately.