Remove the coreapi dependency
The upstream coreapi project is dead, and its support in DRF is slated to be
removed in DRF 3.17. As far as I can tell, we don't use Core API schemas in
any way.

Moreover, `coreapi` imports the `pkg_resources` module, which setuptools is
threatening to remove as soon as this December. So to prevent future
breakage, we should try to get rid of all dependencies that do this.