Documentation: Add Google Analytics tag and cookie consent banner (#9395)
<!-- Raise an issue to propose your change
(https://github.com/cvat-ai/cvat/issues).
It helps to avoid duplication of efforts from multiple independent
contributors.
Discuss your ideas with maintainers to be sure that changes will be
approved and merged.
Read the [Contribution guide](https://docs.cvat.ai/docs/contributing/).
-->

<!-- Provide a general summary of your changes in the Title above -->

### Motivation and context
<!-- Why is this change required? What problem does it solve? If it
fixes an open
issue, please link to the issue here. Describe your changes in detail,
add
screenshots. -->
We would like to gather user statistics for the documentation website.
This PR add Google Analytics script and a cookie banner to comply with
GDPR requirements. This PR is a proposal to discuss.

How it should work in general:

1. When a user comes to the documentation website for the first time, we
must display a cookie banner with options to accept and reject the
analytics cookie.
2. While the banner is present (user yet to make their choice), we must
not inject the tracking script on the page.
3. If user accepts tracking, we can proceed to add the script for
analytics for all documentation pages for this user. The next time, the
same user comes to the website we recognize them via cookie (if they did
not use incognito mode) and we don't display the banner.
4. If user rejects cookie, we don't add analytics script. When such a
user returns, we treat them as a newcomer, and we display the banner
again.
5. We have a new link in the footer that allows calling the banner. If
user selects `Reject`, we clear the previously set cookie and remove the
script from the page. If user selects `Accept` we do the same as on the
step 3.

Some notes about the solution, potential topics for discussion, and
doubts:

- Hugo allows adding Google analytics via configuration file (see [Hugo
docs](https://gohugo.io/methods/site/config/#services)), but this leads
to the script being injected on the website build. This way when a user
comes to the website for the first time, the script is already on the
page. This is not correct. So, the solution is to have a separate
template with all code related to GA tag and the banner. It allows to
control the script injection basing on the user choice about cookies. We
also move the GA tag to Github workflow as a variable. It is used on the
production site. For local development, we have a default tag in the
configuration file (`config.toml`).
- <strike>Following the previous: the GA tag is hard coded in the
template. It may be not good for the code maintenance.</strike>UPDATE:
now the GA tag is hardcoded in the Github workflow only (`docs.yml`).
- The template with cookie banner is called through the footer template
assuming we have footer on every documentation page. Subject for
discussion.
- All CSS are subjects for discussions.
- To avoid analytics pollution when running the website locally, the
cookie banner HTML leverages Hugo native production check. In order to
test this PR locally, you should run **hugo server -e production**. This
will use the default GA tag value from the `config.toml`.
- <strike>The solution does not provide a way to reset the cookies via
website interface. I assume users can do it through the browser tools if
necessary. </strike>UPDATE: the solution add a "Cookie settings" link in
the footer. When user selects the link, cookie banner appears. If user
selects `Reject`, we clear the previously set cookies.
- The solution shows banner for all website visitors regardless of their
location. I don't think we should differentiate them, but it's also a
subject for discussion.

### How has this been tested?
<!-- Please describe in detail how you tested your changes.
Include details of your testing environment, and the tests you ran to
see how your change affects other areas of the code, etc. -->

1. Run `GA_TRACKING_ID="UA-000000000-0" HUGO_ENV=production hugo server
--environment production --bind 0.0.0.0` locally.
2. Open the site

    - Check browser console for any JavaScript errors: should not be any
- Check if GA_TRACKING_ID is set: Open browser dev tools → Console →
type window.GA_TRACKING_ID (should return 'UA-000000000-0'
- Check if element exists: In console type
document.getElementById('cookie-consent-banner'): should return the
element
- Verify localStorage: Check Application tab → Local Storage for
cookieConsent: should be no cookies

3. Select `Accept`: Application => localStorage should show
cookieConsent: "true"
4. Clear the application data, select `Reject`: Application =>
localStorage should show cookieConsent: "rejected"

To test it with network requests: in `injectGAScript` function
(`cookie-consent.js`) comment lines 28-31, in HTML
(`cookie-consent.html`) comment lines 5 and 27. Run the server again
locally with `GA_TRACKING_ID="UA-000000000-0" HUGO_ENV=production hugo
server --environment production --bind 0.0.0.0`.

### Checklist
<!-- Go over all the following points, and put an `x` in all the boxes
that apply.
If an item isn't applicable for some reason, then ~~explicitly
strikethrough~~ the whole
line. If you don't do that, GitHub will show incorrect progress for the
pull request.
If you're unsure about any of these, don't hesitate to ask. We're here
to help! -->
- [x] I submit my changes into the `develop` branch
- [ ] I have created a changelog fragment <!-- see top comment in
CHANGELOG.md -->
- [ ] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] I have linked related issues (see [GitHub docs](

https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword))

### License

- [ ] I submit _my code changes_ under the same [MIT License](
https://github.com/cvat-ai/cvat/blob/develop/LICENSE) that covers the
project.
  Feel free to contact the maintainers if that's a concern.

---------

Co-authored-by: Boris Sekachevw <sekachev.bs@gmail.com>