Merge commit from fork
There was no validation during the import process of a project or task backup to check that the filename specified in the query parameter refers to a TUS-uploaded file belonging to the same user. As a result, if an attacker with a CVAT account and a user role knew the filenames of other users' uploads, they could potentially access and steal data by creating projects or tasks using those files.