Prohibit creating annotations with preset ID
This has no legitimate purpose, since a client cannot predict which IDs are
available ahead of time. Moreover, if the specified ID is already taken, the
client will receive a 500 error (on a PUT request) or a 400 error with a
cryptic message (on a PATCH request).

For symmetry, also prohibit updating and deleting annotations with no ID,
although I believe such operations were previously no-ops anyway.