Prohibit creating annotations with preset ID (#9583)
This has no legitimate purpose, since a client cannot predict which IDs
are available ahead of time. Moreover, if the specified ID is already
taken, the client will receive a 500 error (on a PUT request) or a 400
error with a cryptic message (on a PATCH request).

For symmetry, also prohibit deleting annotations with no ID, although I
believe such deletions were previously no-ops anyway.

I originally wanted to also require IDs on an `action=update` request,
but tests show that a) currently "update" can also work as "create", and
b) it's really easy to accidentally rely on this behavior. So for
compatibility I decided to keep it. Maybe we could revisit this later. I
updated the tests to fix such mistakes, though.