UI: add request types to lambda function calls (#9607)
These are used in the Enterprise version to validate annotation requests
to external lambda functions.

Having an explicit type allows us to verify that only fields that are
appropriate to the particular AR type are included, as well as
disambiguate between AR types if a function supports several of them
(like `init_tracking` and `track`).

This is needed to support agent-based tracker functions.

The code for Nuclio-based functions currently does not use the `type`
field, but eventually I'd like to merge the validation logic between
them and external functions, at which point it will.