Delegate EmailAddress creation to allauth (#9629)
This PR fixes an issue where a user cannot log in if they were invited
to an organization using an email address with uppercase letters, but
then manually registered using the same email address in lowercase.

This issue occurs when the CVAT server is configured with required email
verification (ACCOUNT_EMAIL_VERIFICATION = "mandatory"). There is a
[bug](https://github.com/iMerica/dj-rest-auth/blob/545fd1f4c9ae570216989a8f5a563dd938e038af/dj_rest_auth/serializers.py#L113)
in dj_rest_auth when django allauth version is < 0.62 is used. The old
django allauth versions do not "normalize" emails during registration,
but use register-independent email filtration. That means if an
EmailAddress (from django allauth) is created in another register than
User.email -> dj_rest_auth raises an exception (email is not verified).

---------

Co-authored-by: Roman Donchenko <roman@cvat.ai>