[Documentation] Limit the documentation versions to the 6 ones (#9599)
<!-- Raise an issue to propose your change
(https://github.com/cvat-ai/cvat/issues).
It helps to avoid duplication of efforts from multiple independent
contributors.
Discuss your ideas with maintainers to be sure that changes will be
approved and merged.
Read the [Contribution guide](https://docs.cvat.ai/docs/contributing/).
-->

<!-- Provide a general summary of your changes in the Title above -->

### Motivation and context
<!-- Why is this change required? What problem does it solve? If it
fixes an open
issue, please link to the issue here. Describe your changes in detail,
add
screenshots. -->
Currently, we build the documentation for every repository tag starting
from 1.5.0. This results in large list of the versions which do not
differ too much between themselves. The drop-down menu in the website
overflows the visible area at this point.
This PR limits the number of versions to 6 (changeable in
`build_docs.py`). It also reverses the versions, so in the drop-down
menu on the website, the versions go from the latest to the earliest
top-down.

So far, the main blog had only one link to old version documetation, and
it is changed now. I hope there are no other links to the old versions,
so nothing is broken with this PR.

### How has this been tested?
<!-- Please describe in detail how you tested your changes.
Include details of your testing environment, and the tests you ran to
see how your change affects other areas of the code, etc. -->

- Run the website as for production (see site/README.md for
instructions). Deploy with `python -m http.server -d <...>/public`
- Open the website
- Open the drop-down menu for versions. The menu should contain the six
latest versions.

![image](https://github.com/user-attachments/assets/fb80d8c3-be26-4ebb-8a0b-836dd2a43fe7)


### Checklist
<!-- Go over all the following points, and put an `x` in all the boxes
that apply.
If an item isn't applicable for some reason, then ~~explicitly
strikethrough~~ the whole
line. If you don't do that, GitHub will show incorrect progress for the
pull request.
If you're unsure about any of these, don't hesitate to ask. We're here
to help! -->
- [ ] I submit my changes into the `develop` branch
- [ ] I have created a changelog fragment <!-- see top comment in
CHANGELOG.md -->
- [ ] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] I have linked related issues (see [GitHub docs](

https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword))

### License

- [ ] I submit _my code changes_ under the same [MIT License](
https://github.com/cvat-ai/cvat/blob/develop/LICENSE) that covers the
project.
  Feel free to contact the maintainers if that's a concern.