Increase timeout for waiting background job (#9662)
Timeout increase is required for some test functions that expects background job to be finished. There is a bug in tests that use the same task id, and if test execution is rather slow then a conflict may occur. For example, if one test fails because of an expired timeout during background job waiting, but if the next test uses the same test_id then there is still a job in the worker queue for the old request with the same rq_id, which leads to a 409 error