Remove redundant dispatch_uid parameters
The goal of `dispatch_uid` is to prevent the same signal handler from being
registered more than once if the registration code is run again. However,
the `@receiver` decorators are run during module import, which should only
ever happen once. Therefore, it seems that all of the `dispatch_uid`
parameters in our code are redundant. Remove them to simplify the code.

Ironically, the signal handlers registered in the IAM `register_signals`
function actually _do_ need `dispatch_uid`s, but don't have them. This is
because this function is called from `IAMConfig.ready`, which allegedly [1]
can be run multiple times. This hasn't been a problem yet, but add some UIDs
just in case.

[1]: https://docs.djangoproject.com/en/5.2/topics/signals/#preventing-duplicate-signals