CI: run bandit on all files, not just changed ones (#9726)
This is the final patch in the series that started with #5720. With
this, all linters will be run on all files, thus ensuring that a PR
can't silently add linter warnings in files that it doesn't touch
directly.

In addition, this 1) pins the bandit version, and 2) fixes all remaining
bandit warnings.