Make the server Docker image compatible with the k8s runAsNonRoot setting
While the backend does not actually run as root, Kubernetes is unable to
determine that from the username recorded in the Docker image. Fix it by
using a numeric UID/GID. Also, set the UID explicitly when creating the
user, to make sure it remains stable.

Add the setting itself to the Helm chart, so that we can verify that the
change works.

A client requires this for their deployment, and it seems like a useful
thing to implement in general.

Note that there are several other containers in the Helm chart that are
still incompatible with `runAsNonRoot` (including ones that actually run as
root). I plan to deal with those separately.