Make the frontend container run as a non-root user (#9746)
Continuation of #9743. The frontend container actually does currently
run as root, but fortunately, nginx supplies an alternative Docker
image, which does not. The only difference is that without root
privileges, nginx listens on port 8000 rather than 80, so change the
configuration accordingly.

Also, change the ports in the ingress to their symbolic names to avoid
churn in case we have to change the numbers again.