Remove the tuspy dependency
The version we're depending on (0.2.5) has a dependency on python-future,
which had a CVE filed against it (PythonCharmers/python-future#650). While
the CVE itself seems to be bogus (and in fact, tuspy doesn't actually _use_
python-future), it would still be nice to get rid of this dependency to
placate various security dependency scanners.

tuspy did remove this dependency in version 1.0.2. However, this version
isn't immediately compatible with the various hacks we have applied to marry
it to the API client. Instead of adapting them, I chose to throw tuspy out,
and reimplement the TUS protocol. I think this is better for a few reasons:

* We won't have to keep adapting the hacks every time we need to update the
  dependency.

* The SDK can now be installed in an environment together with other packages
  that have incompatible tuspy requirements.

* The reimplementation turned out to be shorter than the original code.

I also removed the CVAT 2.2 compatibility code that was in `get_offset`,
since we're no longer compatible with those versions for other reasons
anyway (e.g., they don't have `/api/requests`).