Remove URL signing
The `/api/auth/signing` endpoint always returns a 500 code. This is because
`PolicyEnforcer.has_permission` checks `view.detail` as the first step,
which doesn't exist for `SigningView`, as it is not a viewset.

This has been broken at least since 4708b5ec, when `PolicyEnforcer` was
introduced. It could potentially be fixed; however, I see no clear use case
for this feature, plus it has no tests or documentation. Given that it lies
in a security-critical area, I would rather reduce our attack area by
removing it entirely.

This also lets us remove a dependency.