Work around spurious "No module named 'pkg_resources'" errors
Datumaro depends on the pyemd library, which fails to be imported in the
CVAT Docker image, because it uses `pkg_resources` without declaring a
dependency on setuptools (see also wmayner/pyemd#67).

Normally this isn't a problem, because Datumaro imports it lazily (and the
Datumaro component that triggers its actual loading is not used by CVAT).
However, the `current_function_name` function calls `inspect.getouterframes`,
which indirectly causes every lazy-imported module to be loaded (this behavior
might be a Python bug). In particular, this loads pyemd, which raises an
exception.

`current_function_name` is usually called while handling another exception,
so when this problem occurs, that original exception becomes lost.

Work around this by replacing `inspect.getouterframes` with custom code that
does the bare minimum necessary to get the function name.