Clean up cloud provider-related identifiers in the backend code
Currently, the naming pattern is inconsistent and sometimes doesn't match
the official name of the external service. This patch renames such
identifiers, starting from the following official names:

* Amazon S3 (ref: <https://aws.amazon.com/s3/>).

* Azure Blob Storage (ref:
  <https://azure.microsoft.com/en-us/products/storage/blobs>).

* Google Cloud Storage (ref: <https://cloud.google.com/storage?hl=en>). I
  guess technically they just call it "Cloud Storage", but it would be
  ambiguous without Google.

Specifically, I renamed the members of `CloudProviderChoice`, and the
names of the `_CloudStorage` subclasses.

I didn't change the _values_ of `CloudProviderChoice`, because this would be
a breaking API change, and I don't think that's worth doing.

In addition, I:

* Removed the `GOOGLE_DRIVE` member, which has no implementation.
* Used `TextChoices` to give each member a human-readable description.